@@ -10,29 +10,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.11",
+    "@angular/animations": "~14.2.1",
     "@angular/cdk": "13.3.9",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
-    "@angular/forms": "~13.3.11",
+    "@angular/common": "~14.2.1",
+    "@angular/compiler": "~14.2.1",
+    "@angular/core": "~14.2.1",
+    "@angular/forms": "~14.2.1",
     "@angular/material": "13.3.9",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/platform-browser": "~14.2.1",
+    "@angular/platform-browser-dynamic": "~14.2.1",
+    "@angular/router": "~14.2.1",
     "rxjs": "~6.6.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.7",
+    "@angular-devkit/build-angular": "~14.2.2",
     "@angular-eslint/builder": "14.0.4",
     "@angular-eslint/eslint-plugin": "14.0.4",
     "@angular-eslint/eslint-plugin-template": "14.0.4",
     "@angular-eslint/schematics": "14.0.4",
     "@angular-eslint/template-parser": "14.0.4",
     "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.11",
+    "@angular/compiler-cli": "~14.2.1",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "^5.36.2",
